+++ Ranger/Model/PlayerNames.swift
@@ -1,0 +1,22 @@
+//
+//  PlayerNames.swift
+//  Ranger
+//
+//  Created by Geri Borbás on 2020. 01. 22..
+//  Copyright © 2020. Geri Borbás. All rights reserved.
+//
+
+import Foundation
+
+
+extension Model
+{
+    
+    
+    struct PlayerNames: Decodable
+    {
+        
+        
+        let playerNames: [String]
+    }
+}